Feat: add merge_compiled_structures and relocate compilation into rag.flow.extractor
- Move compile_structure_from_text and the supporting _struct_* helpers
  from rag/prompts/generator.py into rag/flow/extractor/extractor.py so
  the structured-knowledge pipeline lives alongside its caller. The
  generator module now re-exports nothing new; existing run_toc_from_text /
  gen_json / split_chunks / INPUT_UTILIZATION are imported from there.

- Add merge_compiled_structures(docs, chat_mdl, embd_mdl, tenant_id, kb_id,
  similarity_threshold=0.9) which is meant to run on the docs returned by
  compile_structure_from_text before they are written to ES:
    Phase 1 (local dedup): group docs by (doc_id, compile_kwd,
      src_name_kwd?, target_name_kwd?), compute pairwise cosine similarity
      over q_<dim>_vec via sklearn.metrics.pairwise.cosine_similarity, and
      for each pair above the threshold ask the LLM via _struct_merge_pair
      whether they are the same logical entity/relation. On a duplicate
      verdict the surviving entry is rebuilt from the merged payload:
      chunk_ids are unioned, the description is re-embedded, src/target
      are forced back to the existing payload's values for relations, and
      the kept entry's id and identity kwds are preserved.
    Phase 2 (ES dedup): for each surviving doc, build the same filter
      condition and KNN-search ES via MatchDenseExpr (topn=1, similarity >=
      threshold). On a hit + duplicate verdict, update the existing ES doc
      by its old id via settings.docStoreConn.update; otherwise insert via
      settings.docStoreConn.insert. Returns
      {"inserted": N, "updated": M, "duplicates_dropped": K}.

- Merge prompts: MERGE_SYSTEM_PROMPT and MERGE_USER_PROMPT are kept
  verbatim; a small MERGE_DECISION_INSTRUCTION is appended so the LLM also
  emits {"duplicated": bool, "merged": <json|null>} via gen_json,
  preserving the user-supplied prompts untouched.

- task_executor.py: import compile_structure_from_text and the new
  merge_compiled_structures from rag.flow.extractor.extractor (no longer
  rag.prompts.generator). knowledge_compilation now reads the structure
  YAML from parser_config["knowledge_compilation"], runs the compile step,
  and then calls merge_compiled_structures. The previously dead toc-style
  branch is replaced by a dedicated kc_thread gated on
  parser_config["knowledge_compilation"] so the new pipeline is opt-in
  via its own flag.
