+++ rag/prompts/generator.py
@@ -22,8 +22,7 @@ from copy import deepcopy
 from typing import Tuple
 from jinja2.sandbox import SandboxedEnvironment
 import json_repair
-import xxhash
-from common.misc_utils import hash_str2int, thread_pool_exec
+from common.misc_utils import hash_str2int
 from rag.nlp import rag_tokenizer
 from rag.prompts.template import load_prompt
 from common.constants import TAG_FLD
@@ -981,524 +980,3 @@ async def multi_queries_gen(chat_mdl, question: str, query:str, missing_infos:li
     return {}
 
 
-# ---------------------------------------------------------------------------
-# Structured-knowledge extraction: list / set / hypergraph
-# ---------------------------------------------------------------------------
-#
-# The configuration shape mirrors the YAML templates used by Hyper-Extract
-# (see D:/git/Hyper-Extract/hyperextract/templates/presets/general/*.yaml and
-# hyperextract/utils/template_engine/parsers/guideline.py). It is delivered to
-# us through ``document.parser_config`` as a JSON string and contains:
-#
-#   type:        "list" | "set" | "hypergraph"      (optional, can be inferred)
-#   output:                                          schema description
-#     # for list/set:
-#     description: <str | list | {lang: str|list}>
-#     fields: [{name, type, description, required?}, ...]
-#     # for hypergraph:
-#     description: ...
-#     entities: {description, fields: [...]}
-#     relations: {description, fields: [...]}
-#   guideline:
-#     # for list/set:
-#     target: <localizable>
-#     rules:  <localizable>
-#     # for hypergraph:
-#     target: <localizable>
-#     rules_for_entities: <localizable>
-#     rules_for_relations: <localizable>
-#     rules_for_time: <localizable, optional>
-#   identifiers:                                     (optional)
-#     entity_id: <field name>                       (used to enumerate known
-#                                                    nodes in stage 2)
-#   options:                                         (optional)
-#     observation_time: <str>                       (substituted into rules)
-#
-# Merging across chunks is intentionally NOT implemented here.
-
-_STRUCT_TYPES = ("list", "set", "hypergraph")
-
-
-def _struct_localize(value, language: str = "en") -> str:
-    """Render multilingual values to a single string (mirrors loader._localize_data)."""
-    if value is None:
-        return ""
-    if isinstance(value, str):
-        return value
-    if isinstance(value, list):
-        return "\n".join(f"{i + 1}. {item}" for i, item in enumerate(value))
-    if isinstance(value, dict):
-        v = value.get(language)
-        if v is None and language != "en":
-            v = value.get("en")
-        if isinstance(v, str):
-            return v
-        if isinstance(v, list):
-            return "\n".join(f"{i + 1}. {item}" for i, item in enumerate(v))
-    return ""
-
-
-def _struct_get(cfg: dict, *keys, default=None):
-    """Case-insensitive lookup against the first matching key."""
-    if not isinstance(cfg, dict):
-        return default
-    for k in keys:
-        if k in cfg:
-            return cfg[k]
-        kl = k.lower()
-        for ck in cfg.keys():
-            if isinstance(ck, str) and ck.lower() == kl:
-                return cfg[ck]
-    return default
-
-
-def _struct_infer_type(parser_config: dict) -> str:
-    explicit = _struct_get(parser_config, "compile_type")
-    if isinstance(explicit, str) and explicit.lower() in _STRUCT_TYPES:
-        return explicit.lower()
-    output = _struct_get(parser_config, "output", default={}) or {}
-    if _struct_get(output, "entities") and _struct_get(output, "relations"):
-        return "hypergraph"
-    return "list"
-
-
-def _struct_render_fields(fields: list, language: str) -> Tuple[str, str]:
-    """Return (bulleted field descriptions, JSON skeleton for one item)."""
-    lines = []
-    skeleton_parts = []
-    for f in fields or []:
-        name = f.get("name", "")
-        ftype = f.get("type", "str")
-        desc = _struct_localize(f.get("description", ""), language)
-        required = f.get("required")
-        req_label = "optional" if required is False else "required"
-        lines.append(f"- {name} ({ftype}, {req_label}): {desc}")
-        if ftype == "list":
-            placeholder = "[<string>, ...]"
-        elif ftype == "int":
-            placeholder = "<int>"
-        elif ftype == "float":
-            placeholder = "<float>"
-        elif ftype == "bool":
-            placeholder = "<true|false>"
-        else:
-            placeholder = "<string>"
-        skeleton_parts.append(f'"{name}": {placeholder}')
-    return "\n".join(lines), "{ " + ", ".join(skeleton_parts) + " }"
-
-
-def _struct_hypergraph_prompts(parser_config: dict, language: str="en") -> Tuple[str, str]:
-    autotype = _struct_get(parser_config, "compile_type", default="graph")
-    guideline = _struct_get(parser_config, "guideline", default={}) or {}
-    output = _struct_get(parser_config, "output", default={}) or {}
-    options = _struct_get(parser_config, "options", default={}) or {}
-
-    target = _struct_localize(_struct_get(guideline, "target"), language)
-    rules_e = _struct_localize(_struct_get(guideline, "rules_for_entities"), language)
-    rules_r = _struct_localize(_struct_get(guideline, "rules_for_relations"), language)
-    rules_t = _struct_localize(_struct_get(guideline, "rules_for_time"), language)
-
-    observation_time = _struct_get(options, "observation_time") or datetime.date.today().isoformat()
-    if rules_t and "{observation_time}" in rules_t:
-        rules_t = rules_t.replace("{observation_time}", observation_time)
-
-    entities_cfg = _struct_get(output, "entities", default={}) or {}
-    relations_cfg = _struct_get(output, "relations", default={}) or {}
-    ent_desc = _struct_localize(_struct_get(entities_cfg, "description"), language)
-    rel_desc = _struct_localize(_struct_get(relations_cfg, "description"), language)
-    ent_fields_text, ent_skel = _struct_render_fields(_struct_get(entities_cfg, "fields", default=[]) or [], language)
-    rel_fields_text, rel_skel = _struct_render_fields(_struct_get(relations_cfg, "fields", default=[]) or [], language)
-
-    node_parts = [f"# Role and Task:\n{target}"] if target else []
-    if rules_e:
-        node_parts.append(f"## Entity Extraction Rules:\n{rules_e}")
-    if ent_desc:
-        node_parts.append(f"## Entity Description:\n{ent_desc}")
-    node_parts.append(f"## Entity Fields:\n{ent_fields_text}")
-    node_parts.append(
-        "## Response Format:\n"
-        "Reply with a single JSON object of the form: "
-        f'{{"items": [{ent_skel}, ...]}}.\n'
-        f"Auto-type: \"{autotype}\". "
-        + ("Items must be unique. " if autotype == "set" else "")
-        + "Return JSON only, no commentary."
-    )
-    node_prompt = "\n\n".join(node_parts)
-
-    if not relations_cfg:
-        return node_prompt, ""
-
-    edge_parts = [f"# Role and Task:\n{target}"] if target else []
-    if rules_r:
-        edge_parts.append(f"## Relation Extraction Rules:\n{rules_r}")
-    if rules_t:
-        edge_parts.append(f"## Time Rules:\n{rules_t}")
-    if rel_desc:
-        edge_parts.append(f"## Relation Description:\n{rel_desc}")
-    edge_parts.append(f"## Relation Fields:\n{rel_fields_text}")
-    edge_parts.append("## Known Entities:\n{known_nodes}")
-    edge_parts.append(
-        "## Response Format:\n"
-        "Reply with a single JSON object of the form: "
-        f'{{"items": [{rel_skel}, ...]}}.\n'
-        "Only create relations between entities listed in 'Known Entities'. "
-        "Return JSON only, no commentary."
-    )
-    edge_prompt = "\n\n".join(edge_parts)
-
-    return node_prompt, edge_prompt
-
-
-def _struct_entity_id_field(parser_config: dict) -> str:
-    identifiers = _struct_get(parser_config, "identifiers", default={}) or {}
-    entity_id = _struct_get(identifiers, "entity_id")
-    if isinstance(entity_id, str) and "{" not in entity_id and entity_id.strip():
-        return entity_id.strip()
-    entities_cfg = _struct_get(_struct_get(parser_config, "output", default={}) or {}, "entities", default={}) or {}
-    for f in _struct_get(entities_cfg, "fields", default=[]) or []:
-        if f.get("required") is not False:
-            return f.get("name", "name")
-    return "name"
-
-
-def _struct_unwrap_items(res) -> list:
-    if res is None:
-        return []
-    if isinstance(res, dict):
-        items = res.get("items")
-        if isinstance(items, list):
-            return [it for it in items if isinstance(it, dict)]
-        return []
-    if isinstance(res, list):
-        return [it for it in res if isinstance(it, dict)]
-    return []
-
-
-async def _struct_extract_hypergraph(text: str, parser_config: dict, chat_mdl, language: str) -> Tuple[list[dict], list[dict]]:
-    node_prompt, edge_prompt_template = _struct_hypergraph_prompts(parser_config, language)
-
-    user_prompt = f"## Source Text:\n{text}\n\n## Output (JSON only):"
-    node_res = await gen_json(node_prompt, user_prompt, chat_mdl, gen_conf={"temperature": 0.1})
-    nodes = _struct_unwrap_items(node_res)
-
-    id_field = _struct_entity_id_field(parser_config)
-    known_keys = []
-    for n in nodes:
-        v = n.get(id_field)
-        if v is None:
-            continue
-        v_str = str(v).strip()
-        if v_str and v_str not in known_keys:
-            known_keys.append(v_str)
-    known_str = "- " + "\n- ".join(known_keys) if known_keys else "(none)"
-
-    edge_prompt = edge_prompt_template.replace("{known_nodes}", known_str)
-    edge_res = await gen_json(edge_prompt, user_prompt, chat_mdl, gen_conf={"temperature": 0.1})
-    edges = _struct_unwrap_items(edge_res)
-
-    return nodes, edges
-
-
-async def _struct_embed(payloads: list[str], embd_mdl) -> list:
-    if not payloads:
-        return []
-    embeddings, _ = await thread_pool_exec(embd_mdl.encode, payloads)
-    return list(embeddings)
-
-
-def _struct_payload_description(payload: dict) -> str:
-    """Concat string values of every non-description field (lists flattened)."""
-    parts: list[str] = []
-    for k, v in payload.items():
-        if isinstance(v, (list, tuple)):
-            for item in v:
-                if item is None:
-                    continue
-                s = str(item).strip()
-                if s:
-                    parts.append(s)
-        else:
-            s = str(v).strip()
-            if s:
-                parts.append(s)
-    return " ".join(parts)
-
-
-def _struct_relation_member_fields(parser_config: dict) -> Tuple:
-    """Return (source_field, target_field) for relation docs, or (None, None).
-
-    Looks at ``identifiers.relation_members`` first (dict form for graph-style
-    configs, e.g. ``{source: source, target: target}``); falls back to the
-    conventional ``source`` / ``target`` field names if both appear in the
-    relation schema.
-    """
-    identifiers = _struct_get(parser_config, "identifiers", default={}) or {}
-    members = _struct_get(identifiers, "relation_members")
-    if isinstance(members, dict):
-        src = members.get("source") or members.get("src")
-        tgt = members.get("target") or members.get("tgt")
-        if src or tgt:
-            return src, tgt
-
-    relations_cfg = _struct_get(
-        _struct_get(parser_config, "output", default={}) or {},
-        "relations",
-        default={},
-    ) or {}
-    field_names = {
-        f.get("name")
-        for f in (_struct_get(relations_cfg, "fields", default=[]) or [])
-        if isinstance(f, dict)
-    }
-    if "source" in field_names and "target" in field_names:
-        return "source", "target"
-    return None, None
-
-
-def _struct_to_es_doc(
-    payload: dict,
-    compile_kwd: str,
-    doc_id: str,
-    chunk_ids: list[str],
-    vec,
-    kind: str,
-    src_field: str | None = None,
-    target_field: str | None = None,
-) -> dict:
-    """Build one ES doc for an extracted entity or relation.
-
-    Args:
-        kind: ``"entity"`` or ``"relation"`` — written to ``structure_kwd``.
-        src_field / target_field: when ``kind == "relation"`` and these field
-            names exist on the payload, the resolved values are written to
-            ``src_name_kwd`` / ``target_name_kwd``.
-    """
-    content_with_weight = json.dumps(payload, ensure_ascii=False)
-    if hasattr(vec, "tolist"):
-        vec_list = vec.tolist()
-    else:
-        vec_list = list(vec)
-    doc_id_str = str(doc_id)
-
-    description = _struct_payload_description(payload)
-
-    content_ltks = rag_tokenizer.tokenize(description) if description else ""
-    content_sm_ltks = rag_tokenizer.fine_grained_tokenize(content_ltks) if content_ltks else ""
-
-    doc = {
-        "content_with_weight": content_with_weight,
-        "compile_kwd": compile_kwd,
-        "structure_kwd": kind,
-        "doc_id": doc_id_str,
-        "chunk_ids": list(chunk_ids or []),
-        "content_ltks": content_ltks,
-        "content_sm_ltks": content_sm_ltks,
-        f"q_{len(vec_list)}_vec": vec_list,
-        "id": xxhash.xxh64(
-            (content_with_weight + doc_id_str).encode("utf-8", "surrogatepass")
-        ).hexdigest(),
-    }
-
-    if kind == "relation":
-        if src_field:
-            src_val = payload.get(src_field)
-            if src_val is not None and str(src_val).strip():
-                doc["src_name_kwd"] = str(src_val).strip()
-        if target_field:
-            tgt_val = payload.get(target_field)
-            if tgt_val is not None and str(tgt_val).strip():
-                doc["target_name_kwd"] = str(tgt_val).strip()
-
-    return doc
-
-
-async def _struct_process_batch(
-    batch: list,
-    batch_idx: int,
-    total: int,
-    chunk_ids: list,
-    autotype: str,
-    parser_config: dict,
-    chat_mdl,
-    embd_mdl,
-    doc_id: str,
-    language: str,
-    callback,
-    semaphore,
-) -> list[dict]:
-    """Process one packed batch end-to-end (extract → embed → ES docs).
-
-    The semaphore (if any) is taken around the entire batch's LLM + embedding
-    work to bound peak concurrency.
-    """
-    if not batch:
-        return []
-
-    batch_ids: list = []
-    batch_segments: list[str] = []
-    for item in batch:
-        for idx, text in item.items():
-            cid = chunk_ids[idx]
-            if cid:
-                batch_ids.append(cid)
-            batch_segments.append(text)
-    combined_text = "\n\n---\n\n".join(batch_segments)
-
-    src_field, target_field = _struct_relation_member_fields(parser_config)
-
-    async def _run() -> list[dict]:
-        # For hypergraph, entity extraction MUST complete before edge extraction
-        # within the same batch, because the edge prompt's {known_nodes}
-        # placeholder is filled from this batch's extracted nodes — see
-        # _struct_extract_hypergraph. Parallelism across batches is fine; the
-        # two stages within one batch are strictly sequential.
-        try:
-            items, relations = await _struct_extract_hypergraph(combined_text, parser_config, chat_mdl, language)
-        except Exception as e:
-            logging.exception(f"compile_structure_from_text: extraction failed for batch {batch_idx}: {e}")
-            return []
-
-        payloads = items + relations
-        kinds = ["entity"] * len(items) + ["relation"] * len(relations)
-        if not payloads:
-            if callback:
-                callback((batch_idx + 1) / total, f"{batch_idx + 1}/{total} batches: 0 items")
-            return []
-
-        embed_inputs = [_struct_payload_description(p) for p in payloads]
-        try:
-            embeddings = await _struct_embed(embed_inputs, embd_mdl)
-        except Exception as e:
-            logging.exception(f"compile_structure_from_text: embedding failed for batch {batch_idx}: {e}")
-            return []
-
-        if len(embeddings) != len(payloads):
-            logging.error(
-                f"compile_structure_from_text: embedding count mismatch ({len(embeddings)} vs {len(payloads)}) for batch {batch_idx}"
-            )
-            return []
-
-        docs = [
-            _struct_to_es_doc(
-                payload, autotype, doc_id, batch_ids, vec, kind,
-                src_field=src_field, target_field=target_field,
-            )
-            for payload, vec, kind in zip(payloads, embeddings, kinds)
-        ]
-
-        if callback:
-            callback((batch_idx + 1) / total, f"{batch_idx + 1}/{total} batches: {len(payloads)} items")
-
-        return docs
-
-    if semaphore is not None:
-        async with semaphore:
-            return await _run()
-    return await _run()
-
-
-async def compile_structure_from_text(
-    chunks: list[dict],
-    parser_config,
-    chat_mdl,
-    embd_mdl,
-    doc_id: str,
-    language: str = "en",
-    callback=None,
-    max_workers: int = 10,
-) -> list[dict]:
-    """Extract list/set/hypergraph structures from text chunks and prepare ES docs.
-
-    Each chunk is processed independently — cross-chunk merging of entities and
-    relations is deferred to a separate pipeline stage and is intentionally not
-    performed here.
-
-    Args:
-        chunks: list of dicts; each must expose ``id`` and ``text`` (a
-            ``content_with_weight`` fallback is also accepted).
-        parser_config: dict already parsed from ``document.parser_config`` or
-            the raw JSON string from the database.
-        chat_mdl: LLMBundle for chat (used via ``gen_json``).
-        embd_mdl: LLMBundle for embeddings (used via ``encode``).
-        doc_id: source document id, embedded into every ES doc.
-        language: language code for resolving multilingual config strings.
-        callback: optional progress callback ``(prog: float, msg: str)``.
-
-    Returns:
-        List of ES-ready dicts shaped as::
-
-            {
-                "content_with_weight": <json>,
-                "compile_kwd": "list" | "set" | "hypergraph",
-                "doc_id": <doc_id>,
-                "chunk_ids": [<chunk_id>, ...],
-                "q_<dim>_vec": [...],
-                "id": <xxhash>,
-            }
-    """
-    if isinstance(parser_config, str):
-        try:
-            parser_config = json.loads(parser_config)
-        except Exception as e:
-            logging.exception(f"compile_structure_from_text: invalid parser_config JSON: {e}")
-            return []
-    if not isinstance(parser_config, dict):
-        logging.error("compile_structure_from_text: parser_config must be a dict or JSON string")
-        return []
-
-    autotype = _struct_infer_type(parser_config)
-    if autotype not in _STRUCT_TYPES:
-        logging.error(f"compile_structure_from_text: unsupported type '{autotype}'")
-        return []
-
-    chunk_ids: list = []
-    chunk_texts: list[str] = []
-    for chunk in chunks:
-        text = chunk.get("text") or chunk.get("content_with_weight") or chunk.get("content") or ""
-        if not isinstance(text, str) or not text.strip():
-            continue
-        chunk_ids.append(chunk.get("id") or chunk.get("chunk_id"))
-        chunk_texts.append(text)
-
-    if not chunk_texts:
-        return []
-
-    node_prompt, edge_prompt = _struct_hypergraph_prompts(parser_config, language)
-    prompt_overhead = max(num_tokens_from_string(node_prompt), num_tokens_from_string(edge_prompt))
-
-    input_budget = int(chat_mdl.max_length * INPUT_UTILIZATION) - prompt_overhead
-    if input_budget < 1024:
-        input_budget = 1024
-
-    batches = split_chunks(chunk_texts, input_budget)
-    total = max(1, len(batches))
-    semaphore = asyncio.Semaphore(max_workers) if max_workers and max_workers > 0 else None
-
-    tasks = [
-        asyncio.create_task(
-            _struct_process_batch(
-                batch, bi, total, chunk_ids, autotype,
-                parser_config, chat_mdl, embd_mdl, doc_id,
-                language, callback, semaphore,
-            )
-        )
-        for bi, batch in enumerate(batches)
-        if batch
-    ]
-
-    if not tasks:
-        return []
-
-    try:
-        batch_results = await asyncio.gather(*tasks, return_exceptions=False)
-    except Exception:
-        for t in tasks:
-            t.cancel()
-        await asyncio.gather(*tasks, return_exceptions=True)
-        raise
-
-    results: list[dict] = []
-    for br in batch_results:
-        if br:
-            results.extend(br)
-    return results
